- make splash screen native (ios & android)
