+++ android/app/src/main/res/drawable-xhdpi/splash.png
