+++ android/app/src/main/res/drawable-xxhdpi/splash.png
