+++ flutter_native_splash.yaml
@@ -1,0 +1,10 @@
+flutter_native_splash:
+  
+  color: "#000000"
+  image: assets/images/appIcon.png
+
+  android_12:
+    color: "#000000"
+    image: assets/images/android12-Icon.png
+
+
