+++ lib/core/routes/routes.dart
@@ -1,3 +1,3 @@
 class Routes {
-  static const String splashScreen = '/'; 
+  static const String home = '/';
 }
