+++ lib/features/home/view/widgets/appLogoAndName.dart
