+++ pubspec.yaml
@@ -37,6 +37,7 @@ dependencies:
   cupertino_icons: ^1.0.6
   flutter_screenutil: ^5.9.0
   flutter_svg: ^2.0.10+1
+  flutter_native_splash: ^2.4.0
 
 dev_dependencies:
   flutter_test:
