+++ web/splash/img/dark-2x.png
