+++ web/splash/img/dark-3x.png
